source/requirements.txt
asyncio
simplematrixbotlib
matrix-nio[e2e]
pyyaml
openai-whisper
crewai
langchain_experimental